requirements_jax.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
arpeggio==2.0.2
astunparse==1.6.3
attrs==23.2.0
av==12.1.0
build==1.2.1
cachecontrol==0.14.0
caliper-reader==0.4.0
causal-conv1d==1.2.2.post1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
cleo==2.1.0
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.8
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.1
decorator==4.4.2
diffusers==0.29.1
dill==0.3.8
distlib==0.3.8
dm-haiku==0.0.12
docstring-parser==0.16
dulwich==0.21.7
einops==0.8.0
etils==1.9.2
executing==2.0.1
farama-notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.13.4
flatbuffers==24.3.25
flax==0.8.5
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.6.0
glfw==2.7.0
google-pasta==0.2.0
grpcio==1.64.1
gymnasium==0.29.1
h5py==3.11.0
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
img2vec-pytorch==1.0.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
installer==0.7.0
jaraco-classes==3.4.0
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jeepney==0.8.0
jinja2==3.1.3
jmp==0.0.4
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
keras==3.4.1
keyring==24.3.1
kiwisolver==1.4.5
libclang==18.1.1
llnl-hatchet==2024.1.2
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.1.6
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.1.1.17
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu12==2.22.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
opt-einsum==3.3.0
optax==0.2.2
optree==0.11.0
orbax-checkpoint==0.5.20
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
pip==24.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry-core==1.9.0
prettytable==3.10.0
proglog==0.1.10
prompt-toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydot==2.0.0
pygments==2.18.0
pyopengl==3.1.7
pyparsing==3.1.2
pyproject-hooks==1.1.0
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.0.0+01cbe5045a
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
ruff==0.4.8
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
secretstorage==3.3.3
setuptools==70.0.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
stable-baselines3==2.3.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.63
termcolor==2.4.0
textx==4.0.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.3
traitlets==5.14.3
transformers==4.41.2
triton==2.3.1
trove-classifiers==2024.5.22
typing-extensions==4.9.0
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.1
uv==0.2.10
virtualenv==20.26.2
wcwidth==0.2.13
werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
